+++ Classes/Public/Manager/OCTPredefined.m
@@ -1,21 +1,17 @@
 //
-//  OCTPredefinedNodes.h
+//  OCTPredefined.m
 //  objcTox
 //
-//  Created by Dmytro Vorobiov on 05/08/15.
+//  Created by Dmytro Vorobiov on 22/08/15.
 //  Copyright (c) 2015 dvor. All rights reserved.
 //
 
-/**
- * A list of nodes from https://wiki.tox.chat/users/nodes
- *
- * See OCTSubmanagerBootstrap for more information
- */
-NSArray *OCTPredefinedNodes()
+#import "OCTPredefined.h"
+
+@implementation OCTPredefined
+
++ (NSArray *)bootstrapNodes
 {
-    /**
-     * Updated 2015-08-05.
-     */
     return @[
         @[
             // sonOfRa, DE
@@ -28,12 +24,6 @@ NSArray *OCTPredefinedNodes()
             @"23.226.230.47",
             @(33445),
             @"A09162D68618E742FFBCA1C2C70385E6679604B2D80EA6E84AD0996A1AC8A074",
-        ],
-        @[
-            // astonex, US
-            @"178.62.125.224",
-            @(33445),
-            @"10B20C49ACBD968D7C80F2E8438F92EA51F189F4E70CFBBB2C2C8C799E97F03E",
         ],
         @[
             // SylvieLorxu, NL
@@ -52,12 +42,6 @@ NSArray *OCTPredefinedNodes()
             @"192.210.149.121",
             @(33445),
             @"F404ABAA1C99A9D37D61AB54898F56793E1DEF8BD46B1038B9D822E8460FAB67",
-        ],
-        @[
-            // Jfreegman, US
-            @"104.219.184.206",
-            @(443),
-            @"8CD087E31C67568103E8C2A28653337E90E6B8EDA0D765D57C6B5172B4F1F04C",
         ],
         @[
             // bunslow, US
@@ -169,3 +153,15 @@ NSArray *OCTPredefinedNodes()
         ],
     ];
 }
+
++ (NSArray *)tox3Servers
+{
+    return @[
+        @[
+            @"utox.org",
+            @"d3154f65d28a5b41a05d4ac7e4b39c6b1c233cc857fb365c56e8392737462a12",
+        ],
+    ];
+}
+
+@end